feat: solution modules task
